+++ extensions/badges/src/createSocialBadge.tsx
@@ -6,7 +6,7 @@ import { Documentation } from "./components/actions.js";
 import { Color, Label, LabelColor, Logo, Message, Style } from "./components/parameters.js";
 import { Input } from "./components/input.js";
 import { colorsForBackground } from "./vendor/badge-maker-color.js";
-import { Badge, LaunchFromSimpleIconsContext } from "./types.js";
+import { Badge, LaunchFromSimpleIconsContext, LaunchFromColorPickerContext } from "./types.js";
 import { codeBlock, encodeBadgeContentParameters } from "./utils.js";
 
 const cache = new Cache();
@@ -23,7 +23,9 @@ const defaultBadge: Badge = {
 };
 const validationFields = ["label", "color", "labelColor", "logoColor"];
 
-export default function Command({ launchContext }: LaunchProps<{ launchContext?: LaunchFromSimpleIconsContext }>) {
+export default function Command({
+  launchContext,
+}: LaunchProps<{ launchContext?: LaunchFromSimpleIconsContext & LaunchFromColorPickerContext }>) {
   const [badge, setBadge] = useCachedState<Badge>("social-badge", defaultBadge);
   const [input, setInput] = useState<{ title: string; value?: string }>({ title: "", value: undefined });
   const [inputValid, setInputValid] = useState(true);
@@ -46,6 +48,11 @@ export default function Command({ launchContext }: LaunchProps<{ launchContext?:
         logo: launchContext.icon.slug,
         logoColor: badge.style === "social" ? undefined : colorsForBackground("#" + launchContext.icon.hex),
       });
+    }
+
+    if (launchContext?.launchFromExtensionName === "color-picker" && launchContext?.hex && launchContext?.field) {
+      console.log(launchContext);
+      setBadge({ ...badge, [launchContext.field]: launchContext.hex.slice(1) });
     }
   }, []);