+++ extensions/badges/src/createStaticBadge.tsx
@@ -4,7 +4,7 @@ import { useCachedState } from "@raycast/utils";
 import { omitBy } from "lodash";
 import { Documentation } from "./components/actions.js";
 import { FieldType, fields } from "./components/parameters.js";
-import { Badge, LaunchFromSimpleIconsContext } from "./types.js";
+import { Badge, LaunchFromSimpleIconsContext, LaunchFromColorPickerContext } from "./types.js";
 import { codeBlock, encodeBadgeContentParameters } from "./utils.js";
 
 const cache = new Cache();
@@ -18,7 +18,9 @@ const defaultBadge: Badge = {
 
 const parameterIds: FieldType[] = ["Label", "Message", "Color", "LabelColor", "Logo", "Style"];
 
-export default function Command({ launchContext }: LaunchProps<{ launchContext?: LaunchFromSimpleIconsContext }>) {
+export default function Command({
+  launchContext,
+}: LaunchProps<{ launchContext?: LaunchFromSimpleIconsContext & LaunchFromColorPickerContext }>) {
   const [badge, setBadge] = useCachedState<Badge>("static-badge", defaultBadge);
   const [input, setInput] = useState<{ title: string; value?: string }>({ title: "", value: undefined });
   const [inputValid, setInputValid] = useState(true);
@@ -36,8 +38,15 @@ export default function Command({ launchContext }: LaunchProps<{ launchContext?:
   };
 
   useEffect(() => {
+    console.log(launchContext);
+
     if (launchContext?.launchFromExtensionName === "simple-icons" && launchContext?.icon) {
       setBadge({ ...badge, $icon: launchContext.icon, logo: launchContext.icon.slug, logoColor: undefined });
+    }
+
+    if (launchContext?.launchFromExtensionName === "color-picker" && launchContext?.hex && launchContext?.field) {
+      console.log(launchContext);
+      setBadge({ ...badge, [launchContext.field]: launchContext.hex.slice(1) });
     }
   }, []);