+++ extensions/color-picker/src/pick-color.ts
@@ -21,13 +21,30 @@ export default async function command(props: PickColorCommandLaunchProps) {
       throw new Error("Failed to format color");
     }
 
-    await Clipboard.copy(hex);
+    if (props.launchContext?.callbackLaunchOptions) {
+      if (props.launchContext?.copyToClipboard) {
+        await Clipboard.copy(hex);
+      }
 
-    await showHUD(`Copied color ${hex} to clipboard`);
+      try {
+        await launchCommand({
+          ...props.launchContext.callbackLaunchOptions,
+          context: {
+            ...props.launchContext.callbackLaunchOptions.context,
+            hex,
+          },
+        });
+      } catch (e) {
+        await showFailureToast(e);
+      }
+    } else {
+      await Clipboard.copy(hex);
+      await showHUD(`Copied color ${hex} to clipboard`);
+    }
+
     try {
       await launchCommand({ name: "menu-bar", type: LaunchType.Background });
     } catch (e) {
-      console.log(e);
       if (!(e instanceof Error && e.message.includes("must be activated"))) {
         await showFailureToast(e);
       }